don't eliminate grey letters which are also yellow
if the correct word is GRAPE and the guess is GREEN, the first E will be
yellow and the second E will be grey. The N will also be grey. But those
two greys must be interpreted very differently! The N is grey because it
doesn't appear at all in the correct word, so we should stop generating
blocks with N labels, while the E is grey because there is only one E in
the correct word, so we should continue generating blocks with E labels.