fix the bug
boardWithLetterBlocksE was using withSimultaneousEvent
boardWithWildBlocksE instead of withBehaviour boardB because the first
completionE occurrence is indeed simultaneous, and thus boardB would not
have the changes yet. unfortunately, the later completionE occurrences
are not simultaneous, so boardWithLetterBlocksE did not trigger at all.

solution: slightly delay the first completionE occurence so it is not
simultaneous.