don't spend time not completing complete words
the code used to show one possible way to complete the word by replacing
the wild blocks with letters, pause for a bit to let the user see the
change, then show a different completion. for rows with no wilds, that
just resulted in the word staying on screen for a while.