allow older GHC versions
I'm struggling with packaging the game so others can use it, so I'm
trying to support ghc-8.8.4 so I can use the asterius compiler to
compile to javascript and post the game online. I assume I will have to
replace the SDL bindings with something which works with the browser,
but I hope I'll be able to reuse everything else.