auto-quit on startup (!)
with the stubs, the program just outputs a lot of text indicating which
functions would have been run. the game loop runs forever (since the
stubbed event loop never returns any event, the game forever waits for
the player to pick a letter), so the program outputs an infinite amount
of text. that's a bit excessive.

now the stubbed event loop emits an infinite amount of quit events, so
the game loop terminates after a single loop. that's no good for a game,
but it's great for comparing a finite amount of text produced with the
Linux executable with a finite amount of text produced by the wasm
executable.