re-dist when reusing previously-build binaries
this was a strange bug:

  1. edit the code to A
  2. run "make build"
  3. edit the code to B
  4. run "make build"
  5. edit the code to A
  6. run "make build"

  expected: tordle-exe.html behaves like A
  actual:   tordle-exe.html behaves like B

the cause is that ahc-cabal noticed that A the tordle-exe pseudo-binary
was already built in the past, and replaced tordle-exe with a cached
copy of that previously-build binary. unfortunately, that cached copy
has a modification date in the past, which defeats Make's change
detection algorithm and thus prevents the subsequent ahc-dist step which
converts the pseudo-binary into the files used by "tordle-exe.html".

the solution is simply to update the modification time of the
pseudo-binary after running ahc-cabal.